fixed a old bug in gal::RenderTarget2D for supporting MSAA.  lots of Session State work.

+++ main/rnd/DummySimulator.h
@@ -48,7 +48,7 @@ public:
 		return true;
 	}
 	
-	void update(float32_t dt)
+	void update(float64_t dt)
 	{
 		if (STATE_RUNNING != _state)
 			return;
